Handle Deterministic Session for ORT and Update Tests
